% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{check_network}
\alias{check_network}
\title{check whether the network is correct}
\usage{
check_network(df_edges, detailed = FALSE)
}
\arguments{
\item{df_edges}{date.frame, data.frame with the definition of edges by the columns `from`, `to`, `weight`, `level`, `set`.}

\item{detailed}{bool, if TRUE, the functiont returns list with detailed results.}
}
\value{
In case of mismatch, returns TRUE.
}
\description{
The function checks whether the created network has one "source" and "exit" and everything is connected.
}
\examples{

network <- create_source(set = c(2,2),
                         weight = c(1,2)) \%>\%
  add_exit()

check_network(network, detailed = TRUE)

}
