% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{flow_coord}
\alias{flow_coord}
\title{Add coordinates to flow}
\usage{
flow_coord(df_flow, df_x, df_y)
}
\arguments{
\item{df_flow}{data.frame with flow generated in `simulate_flow`}

\item{df_x}{data.frame created by function `get_x`}

\item{df_y}{data.frame created by function `get_y`}
}
\value{
data.frame ready for the `animate_simulation` df_text parameter.
}
\description{
Add coordinates to flow
}
