% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{get_coord}
\alias{get_coord}
\title{Get vertices coords}
\usage{
get_coord(df_edges)
}
\arguments{
\item{df_edges}{date.frame, data.frame with the definition of edges by the columns `from`, `to`, `weight`, `level`, `set`.}
}
\value{
data.frame with specified `level`,`vertice`,`set`,`x`,`y`.
}
\description{
The function return data.frame with defined columns `level`,`vertice`,`set`,`x`,`y`. Needed e.g. in plot_network() function.
}
\examples{


network <- create_source(set = c(2,2),
                         weight = c(1,2)) \%>\%
   add_exit()

get_coord(df_edges = network)


}
