% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_x}
\alias{get_x}
\title{Get x coordinates}
\usage{
get_x(df_simulation, x = NULL, side = "left", distance = 1.05)
}
\arguments{
\item{df_simulation}{data.frame created in the `simulate_flow` function.}

\item{x}{numeric vector, it is possible to provide x coordinates manually.}

\item{side}{string, it should be either 'left' or 'right' values to get coordinates to the left or right of the plot.}

\item{distance}{numeric, the distance from the plot.}
}
\value{
data.frame with columns `level` and `x`.
}
\description{
Get x coordinates
}
