% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_y}
\alias{get_y}
\title{Get y coordinates for levels}
\usage{
get_y(df_simulation)
}
\arguments{
\item{df_simulation}{data.frame created in the `simulate_flow` function.}
}
\value{
data.frame with columns `level` and `y`.
}
\description{
Get y coordinates for levels
}
