% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot network from data.frame. Plotting is based on plotly.}
\usage{
plot_network(
  df_edges,
  graph_title = "Network",
  marker_list = list(size = 10),
  edge_shape_color = "rgba(0,100,200,0.4)",
  edge_shape_scale = 1,
  edge_shape_size_normalization = TRUE,
  df_vertices = NULL,
  color = ~paste0(level, "_", set),
  colors = rep(c("#e6194b", "#3cb44b", "#ffe119", "#4363d8", "#f58231", "#911eb4",
    "#46f0f0", "#f032e6", "#bcf60c", "#008080", "#e6beff", "#9a6324", "#800000",
    "#808000", "#ffd8b1", "#000075", "#fabebe", "#808080", "#fffac8", "#000000"), rep =
    10)
)
}
\arguments{
\item{df_edges}{data.frame with the definition of edges by the columns `from`, `to`, `weight`, `level`, `set`.}

\item{graph_title}{string containing the title of the plot. Default is "Network".}

\item{marker_list}{list with elements defining size, shape, etc. of the marker in the plot.}

\item{edge_shape_color}{string of RGBA color for the lines. Default is `rgba(0,100,200,0.4)`.}

\item{edge_shape_scale}{integer describing scale (thickness) of the lines with the smallest weight. Other edges with different weights are proportionally scaled. Default is 1.}

\item{edge_shape_size_normalization}{boolean, TRUE means that the size of the edges is normalize so the maximal size is equivalent to vertice size. This will override `edge_shape_scale` parameter. Default TRUE.}

\item{df_vertices}{data.frame with user defined coordinates `level`, `vertice`, `set`, `x`, `y` for each vertice. Default is NULL. As a example of such data_frame, you could apply get_coord(network) to your network.}

\item{color}{formula, formula for a variable which should be used for coloring. Example `~as.character(set)`. Formula should NOT use `"` enclosure.}

\item{colors}{string, name of a color palette used in plot.}
}
\value{
plotly structure
}
\description{
Plot network from data.frame. Thickness of the edges depends on weight value. Plotting is based on plotly package.
}
\details{
## Parameter `color`:
Color scheme could be based on variable from `df_edges` such as `set`, `level`, etc.

## Parameter `colors`:
Contains name of a color palette as string. There are available following palettes: "Set1", "Set2", "Set3", "Pastel1", "Pastel2", "Paired", "Dark2", and "BrBG".
}
\examples{

network <- create_source(c(2,2),c(1,2)) \%>\%
 add_exit()

##plot network with default options and title: THIS IS NETWORK PLOT.
plot_network(network,
             graph_title = "THIS IS NETWORK PLOT")

## plot network with specified color of lines and marker styles.
plot_network(network,
             marker_list = list(size = 20,
                                color = "green"),
             edge_shape_color = "rgba(255,0,0,0.6)")

## plot vertices with color based on level
plot_network(df_edges = network,
             color = ~as.character(level))

## plot vertices with different color pallete
plot_network(df_edges = network,
             colors = "Paired")

}
